split out pieces with lazy load
